Add support for EBeans and remove JPA + fix tests
