+++ .gitignore
@@ -1,8 +1,10 @@
 logs
 target
+project/.sbtserver
 /.idea
 /.idea_modules
 /.classpath
 /.project
+/.sbtserver
 /.settings
 /RUNNING_PID
