+++ app/views/index.scala.html
@@ -5,7 +5,7 @@
 
     <ul id="ingredients"></ul>
 
-    <form method="POST" action="@routes.Application.addIngredient()">
+    <form method="POST" action="@routes.Ingredient.addIngredient()">
         <input type="text" name="name"/>
         <button>Add Ingredient</button>
     </form>
