+++ conf/application.conf
@@ -340,11 +340,11 @@ db {
   #default.password = ""
 
   # You can expose this datasource via JNDI if needed (Useful for JPA)
-  default.jndiName=DefaultDS
+  # default.jndiName=DefaultDS
 
   # You can turn on SQL logging for any datasource
   # https://www.playframework.com/documentation/latest/Highlights25#Logging-SQL-statements
   #default.logSql=true
 }
 
-jpa.default=defaultPersistenceUnit
+ebean.default = ["models.*"]
