+++ conf/routes
@@ -3,9 +3,9 @@
 # ~~~~
 
 # Home page
-GET     /                           	controllers.Application.index()
-POST    /ingredient                     controllers.Application.addIngredient()
-GET     /ingredients                    controllers.Application.getIngredients()
+GET     /                           	controllers.Ingredient.index()
+POST    /ingredient                     controllers.Ingredient.addIngredient()
+GET     /ingredients                    controllers.Ingredient.getIngredients()
 
 # Map static resources from the /public folder to the /assets URL path
 GET     /assets/*file               controllers.Assets.at(path="/public", file)
