+++ test/IntegrationTest.java
@@ -1,15 +1,9 @@
 import org.junit.*;
 
-import play.mvc.*;
-import play.test.*;
-
-import org.openqa.selenium.htmlunit.HtmlUnitDriver;
-
 import static play.test.Helpers.*;
 import static org.junit.Assert.*;
 import static org.hamcrest.CoreMatchers.*;
 
-import static org.fluentlenium.core.filter.FilterConstructor.*;
 
 public class IntegrationTest {
 
@@ -19,9 +13,8 @@ public class IntegrationTest {
      */
     @Test
     public void test() {
-        running(testServer(3333, fakeApplication(inMemoryDatabase())), HTMLUNIT, browser -> {
-            ((HtmlUnitDriver) browser.getDriver()).setJavascriptEnabled(false);
-            browser.goTo("http://localhost:3333");
+        running(testServer(), HTMLUNIT, browser -> {
+            browser.goTo("/");
             assertThat(browser.pageSource(), containsString("Add Ingredient"));
         });
     }
